+Multiple user support
Multiple users: The root is now a landing page where a contact email can
be submitted, returning the associated data. A new 'account' is made if
the email has not been seen before.

If the session cookie is still good the landing page will redirect to
the WatchList

Router code has been adjusted so that the watch list, product search,
and settings are chidlren of the CenteredTabs component.  This allows
the tabs to appear for all of those pages without also being present on
the landing page.